refactor: extract file system utilities to _fs.py
Moves handle_vscode_settings, merge_json_files, save_init_options, and
load_init_options (plus INIT_OPTIONS_FILE constant) from __init__.py
into the new _fs.py module; re-exports them via __init__.py import.
